// src/app/models/ItemInfo.model.ts
export class ItemInfo {
    id: number;
    name: string;
    arName: string;    
    isActive: boolean;
    DataSourceId:number;

    // constructor(id: number, name: string,  arName: string, isActive: boolean,DataSourceId:number) {
    //     this.id = id;
    //     this.name = name;       
    //     this.isActive = isActive;
    //     this.arName = arName;
    //     this.DataSourceId=DataSourceId;

    // }


}